API

API

Types

ClickHouse.ClickHouseServerException — Type.

ClickHouse server-side exception.

source

ClickHouse.ClickHouseSock — Type.

ClickHouse client socket. Created using connect.

source

Functions

ClickHouse.connect — Function.

Establish a connection to a given ClickHouse instance.

source

ClickHouse.execute — Function.

Execute a DDL query.

source

ClickHouse.insert — Function.

Insert blocks into a table, reading from an iterable. The iterable is expected to yield values of type Dict{Symbol, Any}.

source

ClickHouse.ping — Function.

Send a ping request and wait for the response.

source

ClickHouse.select — Function.

Execute a query, flattening blocks into a single dict of column arrays.

source

ClickHouse.select_channel — Function.

Execute a query, streaming the resulting blocks through a channel.

source

ClickHouse.select_df — Function.

Execute a query, flattening blocks into a dataframe.

source

ClickHouse.select_callback — Function.

Execute a query, invoking a callback for each block.

source